requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
graphviz==0.20.1
